JBEHAVE-1241 Updated configuration of steps instances.  Removed redundant step_monitoring.story
