+++ examples/core/src/main/java/org/jbehave/examples/core/CoreStories.java
@@ -99,8 +99,8 @@ public class CoreStories extends JUnitStories {
         return new InstanceStepsFactory(configuration(),
                 new AndSteps(), new BankAccountSteps(), new BeforeAfterSteps(),
                 new CalendarSteps(), new CompositeSteps(), new CompositeNestedSteps(), new ContextSteps(context), new StepsContextSteps(),
-                new ExamplesTableParametersSteps(),
-                new JsonSteps(), new MetaParametrisationSteps(), new NamedParametersSteps(),
+                new ExamplesTableParametersSteps(), new IgnoringSteps(), new JsonSteps(),
+                new MetaParametrisationSteps(), new NamedAnnotationsSteps(), new NamedParametersSteps(),
                 new ParameterDelimitersSteps(), new ParametrisationByDelimitedNameSteps(), new ParametrisedSteps(),
                 new PendingSteps(), new PriorityMatchingSteps(),
                 new RestartingSteps(), new SandpitSteps(), new SearchSteps(),
